# Time_Management_Project/services/back-main/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Time Management Project",
    "description": "Website for tracking the working hours of employees of an IT company",
    "version": "1.0.0"
  },
  "host": "localhost:9000",
...
  "schemes": "http",
  "paths": {
...
    "/notifications/get_all": {
      "get": {
        "tags": ["Notification"],

        "summary": "Returns a notifications.",
        "description": "Endpoint to get a notifications.",
        "parameters": [
           {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
                 }
                     }
              }
                },
     "/notifications/get/{id_task}": {
      "get": {
        "tags": ["Notification"],
        "summary": "Returns a notification by task.",
        "description": "Endpoint to get a notification by task.",
        "parameters": [
          {
            "name": "id_task",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },           
    "/notifications/get_by_developer/{developer_id}": {
      "get": {
        "tags": ["Notification"],
        "summary": "Returns a notification by developer.",
        "description": "Endpoint to get a notification by developer.",
        "parameters": [
          {
            "name": "developer_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/notifications/get/{task_id}": {
      "get": {
        "tags": ["Notification"],
        "summary": "Returns a notification by task.",
        "description": "Endpoint to get a notification by task.",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/notifications/count/{id_task}": {
      "get": {
        "tags": ["Notification"],
        "summary": "Returns a count by task.",
        "description": "Endpoint to get a count by task.",
        "parameters": [
          {
            "name": "id_task",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }, 
    "/notifications/count_by_developer/{developer_id}": {
      "get": {
        "tags": ["Notification"],
        "summary": "Returns a count by developer id.",
        "description": "Endpoint to get a count by developer id.",
        "parameters": [
          {
            "name": "developer_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }, 
    "/notifications/delete/{notification_id}": {
      "get": {
        "tags": ["Notification"],
        "summary": "Delete a notification.",
        "description": "Endpoint to delete a notification.",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }, 
    "/notifications/put_notification": {
      "post": {
        "tags": ["Notification"],
        "summary": "Put a notification.",
        "description": "Endpoint to put a notification. sender: integer; content: string; date: date-time; taskId: integer",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }, 
    "/update_notification/{notification_id}": {
      "post": {
        "tags": ["Notification"],
        "summary": "Update a notification.",
        "description": "Endpoint to update a notification. content: string",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
           {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/get": {
      "get": {
        "tags": ["Projects"],
        "summary": "Returns projects.",
        "description": "Endpoint to get projects.",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "developer_id",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "type": "integer"
          },
          {
            "name": "pagSize",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/get/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Returns project by id.",
        "description": "Endpoint to get project by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/count": {
      "get": {
        "tags": ["Projects"],
        "summary": "Returns count.",
        "description": "Endpoint to get count.",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "developer_id",
            "in": "query",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/check-for-lead/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Check lead.",
        "description": "Endpoint to check lead.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/put-project": {
      "post": {
        "tags": ["Projects"],
        "summary": "Put project.",
        "description": "Endpoint to put project. title: string, description: string, deadline: date-time, leadId: integer",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/delete/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Delete projects by id.",
        "description": "Endpoint to delete a projects by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/update/{id}": {
      "post": {
        "tags": ["Projects"],
        "summary": "Update project.",
        "description": "Endpoint to update project. title: string, description: string, deadline: date-time, leadId: integer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}